Refactor nav bar on settings screens
Why:
----
prior to this commit, we were using a custom nav bar component for
navigation on the settings screens.

This commit:
----
- updates the nav bar to use React Navigation instead.
