+++ app/bt/AffectedUserFlow/CodeInput.tsx
@@ -54,7 +54,7 @@ const CodeInputScreen = (): JSX.Element => {
   };
 
   const handleOnPressCancel = () => {
-    navigation.navigate(Screens.Settings);
+    navigation.navigate(Screens.More);
   };
 
   const handleOnPressSubmit = async () => {
