+++ app/views/About.js
@@ -13,7 +13,7 @@ import {
 
 import { useStrategyContent } from '../TracingStrategyContext';
 import packageJson from '../../package.json';
-import { NavigationBarWrapper, Typography } from '../components';
+import { Typography } from '../components';
 
 import { useDispatch } from 'react-redux';
 import toggleAllowFeatureFlagsAction from '../store/actions/featureFlags/toggleAllowFeatureFlagsEnabledAction';
@@ -21,7 +21,7 @@ import { Colors, Spacing, Typography as TypographyStyles } from '../styles';
 
 const CLICKS_TO_ENABLE_FEATURE_FLAGS = 10;
 
-export const AboutScreen = ({ navigation }) => {
+export const AboutScreen = () => {
   const dispatch = useDispatch();
 
   const [clickCount, setClickCount] = useState(0);
@@ -37,54 +37,46 @@ export const AboutScreen = ({ navigation }) => {
   const { t } = useTranslation();
   const { StrategyCopy } = useStrategyContent();
 
-  const backToMain = () => {
-    navigation.goBack();
-  };
-
   return (
-    <NavigationBarWrapper
-      title={t('screen_titles.about')}
-      onBackPress={backToMain}>
-      <ScrollView
-        contentContainerStyle={styles.contentContainer}
-        alwaysBounceVertical={false}>
-        <TouchableWithoutFeedback onPress={incrementClickCount}>
-          <View>
-            <Typography use='headline2' style={styles.heading}>
-              {StrategyCopy.aboutHeader}
-            </Typography>
-          </View>
-        </TouchableWithoutFeedback>
-        <Typography use='body2'>{t('label.about_para')}</Typography>
-        <Typography
-          style={styles.hyperlink}
-          onPress={() => {
-            Linking.openURL('https://pathcheck.org/');
-          }}>
-          <Text>{'pathcheck.org'}</Text>
-        </Typography>
-
-        <View style={styles.rowContainer}>
-          <View style={styles.row}>
-            <Typography style={styles.aboutSectionParaLabel}>
-              {t('about.version')}
-            </Typography>
-
-            <Typography style={styles.aboutSectionParaContent}>
-              {packageJson.version}
-            </Typography>
-          </View>
-          <View style={styles.row}>
-            <Typography style={styles.aboutSectionParaLabel}>
-              {t('about.operating_system_abbr')}
-            </Typography>
-            <Typography style={styles.aboutSectionParaContent}>
-              {Platform.OS + ' v' + Platform.Version}
-            </Typography>
-          </View>
+    <ScrollView
+      contentContainerStyle={styles.contentContainer}
+      alwaysBounceVertical={false}>
+      <TouchableWithoutFeedback onPress={incrementClickCount}>
+        <View>
+          <Typography use='headline2' style={styles.heading}>
+            {StrategyCopy.aboutHeader}
+          </Typography>
         </View>
-      </ScrollView>
-    </NavigationBarWrapper>
+      </TouchableWithoutFeedback>
+      <Typography use='body2'>{t('label.about_para')}</Typography>
+      <Typography
+        style={styles.hyperlink}
+        onPress={() => {
+          Linking.openURL('https://pathcheck.org/');
+        }}>
+        <Text>{'pathcheck.org'}</Text>
+      </Typography>
+
+      <View style={styles.rowContainer}>
+        <View style={styles.row}>
+          <Typography style={styles.aboutSectionParaLabel}>
+            {t('about.version')}
+          </Typography>
+
+          <Typography style={styles.aboutSectionParaContent}>
+            {packageJson.version}
+          </Typography>
+        </View>
+        <View style={styles.row}>
+          <Typography style={styles.aboutSectionParaLabel}>
+            {t('about.operating_system_abbr')}
+          </Typography>
+          <Typography style={styles.aboutSectionParaContent}>
+            {Platform.OS + ' v' + Platform.Version}
+          </Typography>
+        </View>
+      </View>
+    </ScrollView>
   );
 };
