+++ app/views/More/ENDebugMenu.tsx
@@ -10,7 +10,6 @@ import {
   ActivityIndicator,
 } from 'react-native';
 
-import { NavigationBarWrapper } from '../../components/NavigationBarWrapper';
 import { Typography } from '../../components/Typography';
 import { BTNativeModule } from '../../bt';
 import { NavigationProp, Screens } from '../../navigation';
@@ -37,10 +36,6 @@ const ENDebugMenu = ({ navigation }: ENDebugMenuProps): JSX.Element => {
       BackHandler.removeEventListener('hardwareBackPress', handleBackPress);
     };
   }, [navigation]);
-
-  const backToSettings = () => {
-    navigation.goBack();
-  };
 
   const showErrorAlert = (errorString: string) => {
     Alert.alert('Error', errorString, [{ text: 'OK' }], {
@@ -109,7 +104,7 @@ const ENDebugMenu = ({ navigation }: ENDebugMenuProps): JSX.Element => {
   };
 
   return (
-    <NavigationBarWrapper title={'EN Debug Menu'} onBackPress={backToSettings}>
+    <>
       {loading ? (
         <View style={{ flex: 1, justifyContent: 'center' }}>
           <ActivityIndicator size={'large'} />
@@ -178,7 +173,7 @@ const ENDebugMenu = ({ navigation }: ENDebugMenuProps): JSX.Element => {
           </View>
         </ScrollView>
       )}
-    </NavigationBarWrapper>
+    </>
   );
 };
